Use TypeScript for tool call comparison data
