+++ website/scripts/compute-llm-call-tokens.ts
@@ -1,9 +1,9 @@
-#!/usr/bin/env node
+#!/usr/bin/env bun
 /**
  * One-shot helper: compute the GPT-5-family / o200k_base token count
  * of the shared Responses API request bodies used by the sandbox page.
  *
- *   node scripts/compute-llm-call-tokens.mjs
+ *   bun scripts/compute-llm-call-tokens.ts
  *
  * `gpt-tokenizer` is a devDependency for this maintenance script only;
  * it is not imported by the website runtime.
@@ -15,23 +15,27 @@ import {
   buildLlmRequest,
   LLM_CALL_TOKENS,
   TOOL_CALL_FLOWS,
-} from "../src/lib/tool-call-comparison.mjs";
+  type ToolFlow,
+  type ToolFlowKey,
+} from "../src/lib/tool-call-comparison";
 
-const tokens = (obj) => encode(JSON.stringify(obj)).length;
+const tokens = (obj: unknown) => encode(JSON.stringify(obj)).length;
 
 const computed = Object.fromEntries(
-  Object.entries(TOOL_CALL_FLOWS).map(([flowKey, flow]) => [
-    flowKey,
-    {
-      in: flow.steps.map((_, i) => tokens(buildLlmRequest(flowKey, i))),
-      out: flow.steps.map((_, i) =>
-        tokens(buildFunctionCallOutput(flowKey, i)),
-      ),
-    },
-  ]),
-);
+  (Object.entries(TOOL_CALL_FLOWS) as [ToolFlowKey, ToolFlow][]).map(
+    ([flowKey, flow]) => [
+      flowKey,
+      {
+        in: flow.steps.map((_, i) => tokens(buildLlmRequest(flowKey, i))),
+        out: flow.steps.map((_, i) =>
+          tokens(buildFunctionCallOutput(flowKey, i)),
+        ),
+      },
+    ],
+  ),
+) as Record<ToolFlowKey, { in: number[]; out: number[] }>;
 
-const fmt = (label, ins, outs) => {
+const fmt = (label: string, ins: number[], outs: number[]) => {
   console.log(`== ${label} ==`);
   ins.forEach((tIn, i) => {
     console.log(
@@ -45,13 +49,16 @@ const fmt = (label, ins, outs) => {
   );
 };
 
-for (const [flowKey, flow] of Object.entries(TOOL_CALL_FLOWS)) {
+for (const [flowKey, flow] of Object.entries(TOOL_CALL_FLOWS) as [
+  ToolFlowKey,
+  ToolFlow,
+][]) {
   if (flowKey !== "bash") console.log();
   fmt(`${flow.label} flow`, computed[flowKey].in, computed[flowKey].out);
 }
 
 console.log(
-  `\nFor src/lib/tool-call-comparison.mjs LLM_CALL_TOKENS:\n  bash:   { in: [${computed.bash.in.join(", ")}], out: [${computed.bash.out.join(", ")}] },\n  goccia: { in: [${computed.goccia.in.join(", ")}], out: [${computed.goccia.out.join(", ")}] },`,
+  `\nFor src/lib/tool-call-comparison.ts LLM_CALL_TOKENS:\n  bash:   { in: [${computed.bash.in.join(", ")}], out: [${computed.bash.out.join(", ")}] },\n  goccia: { in: [${computed.goccia.in.join(", ")}], out: [${computed.goccia.out.join(", ")}] },`,
 );
 
 if (JSON.stringify(computed) !== JSON.stringify(LLM_CALL_TOKENS)) {
